+++ src/landing/header/headerComponents/HeaderButtons.jsx
@@ -1,0 +1,27 @@
+import React from "react";
+import { motion } from "framer-motion";
+import styles from "../Header.module.css";
+
+const HeaderButtons = ({ openLogin, openRegister, isHovered, setIsHovered }) => {
+    return (
+        <div className={styles.headerBtns}>
+            <button
+                className={styles.btnLogin}
+                onClick={openLogin}
+                onMouseEnter={() => setIsHovered(true)}
+                onMouseLeave={() => setIsHovered(false)}
+            >
+                <motion.div
+                    animate={isHovered ? { rotate: [0, -10, 10, -10, 0] } : { rotate: 0 }}
+                    transition={{ duration: 0.5, ease: "easeInOut" }}
+                />
+                Login
+            </button>
+            <button className={styles.btnSignup} onClick={openRegister}>
+                Sign up for Free
+            </button>
+        </div>
+    );
+};
+
+export default HeaderButtons;
